+++ libnavui-base/src/main/java/com/mapbox/navigation/ui/base/lifecycle/UIComponent.kt
@@ -5,14 +5,16 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 /**
- * Using the [UIComponent] gives you access to a [coroutineScope]. All coroutines that you
+ * Using the [UIComponent] gives you access to a [coroutineScope] which uses
+ * `Dispatchers.Main.immediate`. All coroutines that you
  * launch inside onAttached will be canceled when the observer is detached. You do not need to
  * implement onDetached for your flowable components.
  * @property coroutineScope Defines a scope for new coroutine
@@ -28,7 +30,7 @@ open class UIComponent : MapboxNavigationObserver {
      */
     @CallSuper
     override fun onAttached(mapboxNavigation: MapboxNavigation) {
-        coroutineScope = MainScope()
+        coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate)
     }
 
     /**